Extract the iterator functions
