+++ scanner/src/main/kotlin/voice/app/scanner/CoverExtractor.kt
@@ -2,45 +2,14 @@ package voice.app.scanner
 
 import android.content.Context
 import android.net.Uri
-import androidx.media3.common.Format
 import androidx.media3.common.MediaItem
-import androidx.media3.common.Metadata
-import androidx.media3.common.TrackGroup
 import androidx.media3.exoplayer.MetadataRetriever
-import androidx.media3.exoplayer.source.TrackGroupArray
 import androidx.media3.extractor.metadata.flac.PictureFrame
 import androidx.media3.extractor.metadata.id3.ApicFrame
 import kotlinx.coroutines.guava.await
 import voice.logging.core.Logger
 import java.io.File
 import javax.inject.Inject
-
-operator fun TrackGroupArray.iterator(): Iterator<TrackGroup> = object : Iterator<TrackGroup> {
-  private var index = 0
-  override fun hasNext(): Boolean = index < length
-  override fun next(): TrackGroup {
-    if (!hasNext()) throw NoSuchElementException()
-    return get(index++)
-  }
-}
-
-operator fun TrackGroup.iterator() = object : Iterator<Format> {
-  private var index = 0
-  override fun hasNext(): Boolean = index < length
-  override fun next(): Format {
-    if (!hasNext()) throw NoSuchElementException()
-    return getFormat(index++)
-  }
-}
-
-operator fun Metadata.iterator() = object : Iterator<Metadata.Entry> {
-  private var index = 0
-  override fun hasNext(): Boolean = index < length()
-  override fun next(): Metadata.Entry {
-    if (!hasNext()) throw NoSuchElementException()
-    return get(index++)
-  }
-}
 
 class CoverExtractor @Inject constructor(private val context: Context) {
